add more Attach trimming setting
