+++ Notifications.Wpf/Notifications.Wpf.nuspec
@@ -2,7 +2,7 @@
 <package >
   <metadata>
     <id>Notification.WPF</id>
-    <version>1.0.4.0</version>
+    <version>1.0.4.1</version>
     <title>Notification.WPF</title>
     <authors>APlatonenkov</authors>
 	  <owners>https://github.com/Platonenkov</owners>
